cronjob/requirements.txt
# Python dependencies for cron scripts
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# embedding_sync.py dependencies
psycopg[binary]>=3.1
pgvector>=0.2
langchain-openai>=0.1
openai>=1.0